+++ WorkoutManager/WorkoutManager/Config/AppDelegate.swift
@@ -22,7 +22,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     lazy var persistentContainer: NSPersistentContainer = {
         
             let container = NSPersistentContainer(name: "Model")
-            container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+            container.loadPersistentStores(completionHandler: { (_, error) in
                 if let error = error as NSError? {
                     fatalError("Unresolved error \(error), \(error.userInfo)")
                 }
