+++ WorkoutManager/WorkoutManager/Controllers/UserViewController.swift
@@ -2,7 +2,7 @@ import UIKit
 
 class UserViewController: UIViewController {
     
-    private var userUser: User?
+    private var userUser: UserModel?
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -12,7 +12,7 @@ class UserViewController: UIViewController {
     
     func getUserDetails() {
         let url = Constants.baseURL?.appendingPathComponent("userprofile/")
-        URLSession.shared.makeRequest(url: url,method: .get, returnModel:User.self, completion: {[weak self]result in
+        URLSession.shared.makeRequest(url: url,method: .get, returnModel:UserModel.self, completion: {[weak self]result in
 
             switch result {
             case .success(let userArray):
