+++ WorkoutManager/WorkoutManager/Controllers/WorkoutManagerViewController.swift
@@ -1,21 +1,83 @@
+//
+//  ViewController.swift
+//  WorkoutManager
+//
+//  Created by Simekani Mabambi on 2022/03/09.
+//
+
 import UIKit
 
 class WorkoutManagerViewController: UIViewController {
-
+    
     private var workouts: WorkoutManager?
     
-    @IBOutlet private weak var workoutPlans: UICollectionView!
+    @IBOutlet var workoutPlans : UICollectionView!
+
+    private func configureCollectionView() -> UICollectionViewFlowLayout {
+        let layout = UICollectionViewFlowLayout()
+        layout.itemSize = CGSize(width: 230, height: 370)
+        layout.scrollDirection = .horizontal
+        layout.minimumLineSpacing = CGFloat(45)
+        return layout
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.getWorkoutPlans()
         
-        workoutPlans.register(WorkoutsCollectionViewCell.nib(),
-                              forCellWithReuseIdentifier: WorkoutsCollectionViewCell.indetifier)
+        workoutPlans.register(WorkoutPlanCollectionViewCell.nib(),
+                              forCellWithReuseIdentifier: WorkoutPlanCollectionViewCell.identifier)
         
-        self.workoutPlans.delegate = self
-        self.workoutPlans.dataSource = self
+        workoutPlans.collectionViewLayout = self.configureCollectionView()
+        workoutPlans.delegate = self
+        workoutPlans.dataSource = self
+        
+        getWorkoutPlans()
+
     }
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
+
+extension WorkoutManagerViewController: UICollectionViewDelegate, UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        // swiftlint:disable force_cast
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: WorkoutPlanCollectionViewCell.identifier, for: indexPath) as! WorkoutPlanCollectionViewCell
+        
+        guard let finalWorkouts = self.workouts?.workoutPlans  else {return cell}
+        
+        cell.setCellProperties(image: UIImage(named: "workout1")!, label: (finalWorkouts[indexPath.row].name))
+        // swiftlint:enable force_cast
+        return cell
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        collectionView.deselectItem(at: indexPath, animated: true)
+        
+        print("You selescted an item")
+    }
+    
+    func collectionView(_ collectionView: UICollectionView,numberOfItemsInSection section: Int) -> Int {
+        guard let finalWorkouts = self.workouts else {return 0}
+        return finalWorkouts.workoutPlans.count
+    }
+}
+
+extension WorkoutManagerViewController: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: 230, height: 370)
+    }
+}
+
+extension WorkoutManagerViewController {
     
     func getWorkoutPlans() {
         let url = Constants.baseURL?.appendingPathComponent("workout/")
@@ -32,53 +94,5 @@ class WorkoutManagerViewController: UIViewController {
                 print(error)
             }})
     }
-
-}
-
-extension WorkoutManagerViewController : UICollectionViewDataSource, UICollectionViewDelegate {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        
-        guard let finalWorkouts = self.workouts else {return 0}
-        
-        return finalWorkouts.workoutPlans.count
-        
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        // swiftlint:disable force_cast
-        let cell = workoutPlans.dequeueReusableCell(withReuseIdentifier: WorkoutsCollectionViewCell.indetifier, for: indexPath) as! WorkoutsCollectionViewCell
-        // swiftlint:enable force_cast
-        guard let finalWorkouts = workouts else {
-            return cell
-        }
-        
-        cell.setCellAttributes(image: UIImage(named: "NewLogo")!, status: "Active")
-        
-        return cell
-    }
     
 }
-
-// extension WorkoutManagerViewController: UITableViewDelegate, UITableViewDataSource {
-//
-//    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-//        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-//
-//        guard let finalWorkouts = workouts else {
-//            cell.textLabel?.text = "No Workout Plans"
-//            return cell
-//        }
-//
-//        cell.textLabel?.text = finalWorkouts.workoutPlans[indexPath.row].name
-//
-//        return cell
-//    }
-//
-//    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-//
-//        guard let finalWorkouts = self.workouts else {
-//            return 0
-//        }
-//         return finalWorkouts.workoutPlans.count
-//    }
